refactor some essentially global PKI functions into pki module
