+++ adapter/ph/src/km_cert_exchange.rs
@@ -17,19 +17,16 @@
 //!
 //!
 
-use openssl::pkey::PKey;
 use openssl::x509::X509;
-use std::fs;
 use std::path::Path;
 use tracing::error;
 use zerocopy::byteorder::network_endian::*;
 use zerocopy::{FromBytes, Immutable, IntoBytes, KnownLayout, Unaligned};
 
-use crate::km_noise::NOISE_KEY_LEN;
-use crate::logging::targets::KEY_MGMT;
 
-const PEM_BEGIN_CERTIFICATE: &str = "-----BEGIN CERTIFICATE-----";
-const PEM_END_CERTIFICATE: &str = "-----END CERTIFICATE-----";
+use crate::logging::targets::KEY_MGMT;
+use crate::pki::{load_cert, ParseError};
+
 
 #[derive(Debug)]
 pub enum CertExchangeError {
@@ -43,13 +40,6 @@ pub enum CertExchangeError {
     KeyMismatchError,
 }
 
-#[derive(Debug)]
-pub enum ParseError {
-    PEMCertNotFound,
-    PEMFormatError,
-    KeyError,
-    IOError(#[allow(dead_code)] std::io::Error),
-}
 
 #[derive(FromBytes, IntoBytes, Immutable, KnownLayout, Unaligned)]
 #[repr(packed)]
@@ -208,120 +198,15 @@ impl KmCertExchange {
     }
 }
 
-/// Look for first instance of "-----BEGIN CERTIFICATE-----" and return that up to and
-/// including the "-----END CERTIFICATE-----" line.
-///
-/// - `textdata` is the textual representation of a certificate that must include the PEM
-///    encoded certificate in it somewhere.  Typically, a certificate file has the PEM data
-///    at the end.
-///
-/// Returns a copy of the PEM data from the intput without trying to parse it.
-fn extract_cert_pem_data(textdata: &str) -> Result<String, ParseError> {
-    let mut pemdata = String::new();
-    let mut reading = false;
-    for line in textdata.lines() {
-        if reading {
-            pemdata.push_str(line);
-            pemdata.push_str("\n");
-            if line.starts_with(PEM_END_CERTIFICATE) {
-                return Ok(pemdata);
-            }
-        } else if line.starts_with(PEM_BEGIN_CERTIFICATE) {
-            reading = true;
-            pemdata.push_str(line);
-            pemdata.push_str("\n");
-        }
-    }
-    if !reading {
-        Err(ParseError::PEMCertNotFound)
-    } else {
-        Err(ParseError::PEMFormatError)
-    }
-}
 
-/// Load a certificate from a file.
-pub fn load_cert(path: &Path) -> Result<X509, ParseError> {
-    let contents = match fs::read_to_string(path) {
-        Ok(c) => c,
-        Err(e) => return Err(ParseError::IOError(e)),
-    };
-    let cert_pem_data = extract_cert_pem_data(&contents)?;
-    match X509::from_pem(cert_pem_data.as_bytes()) {
-        Ok(cert) => Ok(cert),
-        Err(e) => {
-            error!(target: KEY_MGMT, "error constructing cert from PEM data: {e}");
-            Err(ParseError::PEMFormatError)
-        }
-    }
-}
 
-/// Load a private X22519 key from a PEM file
-pub fn load_private_key(path: &Path) -> Result<[u8; NOISE_KEY_LEN], ParseError> {
-    let contents = match fs::read_to_string(path) {
-        Ok(c) => c,
-        Err(e) => return Err(ParseError::IOError(e)),
-    };
-
-    let pk = match PKey::private_key_from_pem(&contents.as_bytes()) {
-        Ok(k) => k,
-        Err(e) => {
-            error!(target: KEY_MGMT, "error reading key from PEM data: {e}");
-            return Err(ParseError::PEMFormatError);
-        }
-    };
-
-    match pk.raw_private_key() {
-        Ok(k) => {
-            let mut key = [0u8; NOISE_KEY_LEN];
-            key.copy_from_slice(&k);
-            Ok(key)
-        }
-        Err(e) => {
-            error!(target: KEY_MGMT, "error extracting raw key: {e}");
-            Err(ParseError::KeyError)
-        }
-    }
-}
-
-pub fn load_public_key(path: &Path) -> Result<[u8; NOISE_KEY_LEN], ParseError> {
-    let contents = match fs::read_to_string(path) {
-        Ok(c) => c,
-        Err(e) => return Err(ParseError::IOError(e)),
-    };
-
-    let pk = match PKey::public_key_from_pem(&contents.as_bytes()) {
-        Ok(k) => k,
-        Err(e) => {
-            error!(target: KEY_MGMT, "error reading key from PEM data: {e}");
-            return Err(ParseError::PEMFormatError);
-        }
-    };
-
-    match pk.raw_public_key() {
-        Ok(k) => {
-            if k.len() != NOISE_KEY_LEN {
-                error!(
-                    target: KEY_MGMT,
-                    "public key in cert is incorrect length (got {} bytes, expected {NOISE_KEY_LEN})",
-                    k.len(),
-                );
-                return Err(ParseError::KeyError);
-            }
-            let key = <[u8; NOISE_KEY_LEN]>::try_from(k).unwrap();
-            Ok(key)
-        }
-        Err(e) => {
-            error!(target: KEY_MGMT, "error extracting raw key: {e}");
-            Err(ParseError::KeyError)
-        }
-    }
-}
 
 #[cfg(test)]
 mod test {
     use super::*;
-    use crate::km_noise::{derive_public_key, NOISE_KEY_LEN};
+    use crate::km_noise::derive_public_key;
     use crate::km_testdata::test::*;
+    use crate::pki::NOISE_KEY_LEN;
     use base64::prelude::*;
     use bytes::BytesMut;
